+++ lib/xvm-helper/Cargo.toml
@@ -8,8 +8,6 @@ edition = "2021"
 ink_env = { version = "3.4.0", default-features = false }
 ink_lang = { version = "3.4.0", default-features = false }
 ink_prelude = { version = "3.4.0", default-features = false }
-hex-literal = "0.3"
-ethabi = { git = "https://github.com/akru/ethabi", default-features = false }
 
 scale = { package = "parity-scale-codec", version = "3", default-features = false, features = ["derive"] }
 scale-info = { version = "2", default-features = false, features = ["derive"], optional = true }
