+++ xvm-sdk-contracts/psp22-wrapper/lib.rs
@@ -3,18 +3,29 @@
 
 #[openbrush::contract]
 pub mod my_psp22 {
-    use ink_prelude::{
-        vec::Vec,
+    use ethabi::{
+        ethereum_types::{
+            H160,
+            U256,
+        },
+        Token,
     };
+    use hex_literal::hex;
+    use ink_lang::ToAccountId;
+    use ink_prelude::vec::Vec;
     use ink_storage::traits::SpreadAllocate;
     use openbrush::{
-        contracts::psp22::extensions::metadata::*,
+        contracts::{
+            psp22::extensions::metadata::*,
+            traits::psp22::PSP22Ref,
+        },
         traits::Storage,
-        contracts::traits::psp22::PSP22Ref,
     };
     use xvm_sdk_helper::*;
     use xvm_sdk_psp22_controller::Psp22Ref;
-    use ink_lang::ToAccountId;
+
+    const EVM_ID: u8 = 0x0F;
+    const TRANSFER_SELECTOR: [u8; 4] = hex!["a9059cbb"];
 
     #[ink(storage)]
     #[derive(Default, SpreadAllocate, Storage)]
@@ -32,7 +43,11 @@ pub mod my_psp22 {
 
     impl PSP22Wrapper {
         #[ink(constructor)]
-        pub fn new(version: u32, psp22_controller_hash: Hash, evm_contract_address: [u8; 20]) -> Self {
+        pub fn new(
+            version: u32,
+            psp22_controller_hash: Hash,
+            evm_contract_address: [u8; 20],
+        ) -> Self {
             ink_lang::codegen::initialize_contract(|instance: &mut Self| {
                 instance.metadata.name = Some("Wrapped PSP22".as_bytes().to_vec());
                 instance.metadata.symbol = Some("WPSP22".as_bytes().to_vec());
@@ -45,9 +60,12 @@ pub mod my_psp22 {
                     .salt_bytes(salt)
                     .instantiate()
                     .unwrap_or_else(|error| {
-                        panic!("failed at instantiating the psp22 controller contract: {:?}", error)
+                        panic!(
+                            "failed at instantiating the psp22 controller contract: {:?}",
+                            error
+                        )
                     });
-              instance.psp22_controller = psp22.to_account_id();
+                instance.psp22_controller = psp22.to_account_id();
             })
         }
 
@@ -60,7 +78,7 @@ pub mod my_psp22 {
         pub fn deposit(&mut self, amount: Balance) -> Result<(), PSP22Error> {
             let caller = self.env().caller();
             let contract = self.env().account_id();
-            XvmErc20::transfer(self.evm_address, contract, amount, Vec::new())
+            Self::_transfer(self.evm_address, contract, amount, Vec::new())
                 .map_err(|_| PSP22Error::Custom("transfer failed".as_bytes().to_vec()))?;
             self._mint_to(caller, amount)
         }
@@ -70,6 +88,34 @@ pub mod my_psp22 {
             let caller = self.env().caller();
             self._burn_from(caller, amount)?;
             PSP22Ref::transfer(&mut self.psp22_controller, caller, amount, Vec::new())
+        }
+
+        fn _transfer(
+            evm_contract_address: [u8; 20],
+            to: AccountId,
+            value: Balance,
+            _data: Vec<u8>,
+        ) -> Result<(), XvmError> {
+            let encoded_input = Self::_transfer_encode(Self::_h160(&to), value.into());
+            Xvm::xvm_call(
+                EVM_ID,
+                Vec::from(evm_contract_address.as_ref()),
+                encoded_input,
+            )
+        }
+
+        fn _transfer_encode(to: H160, value: U256) -> Vec<u8> {
+            let mut encoded = TRANSFER_SELECTOR.to_vec();
+            let input = [Token::Address(to), Token::Uint(value)];
+            encoded.extend(&ethabi::encode(&input));
+            encoded
+        }
+
+        fn _h160(from: &AccountId) -> H160 {
+            let mut dest: H160 = [0; 20].into();
+            dest.as_bytes_mut()
+                .copy_from_slice(&<ink_env::AccountId as AsRef<[u8]>>::as_ref(from)[..20]);
+            dest
         }
     }
 }
